Refactor mainnet and testnet functionalities into separate webservers
